requirements.txt
streamlit
pillow
pytesseract
google-generativeai
